Exit early if json_file isn't found

+++ tasks/install-ert/configure-ert.sh
@@ -3,6 +3,10 @@ set -e
 
 json_file="json_file/ert.json"
 
+if [[ ! -f ${json_file} ]]; then
+  echo "Error: cant find file=[${json_file}]"
+  exit 1
+fi
 
 # Setup OM Tool
 sudo cp tool-om/om-linux /usr/local/bin
@@ -26,11 +30,6 @@ my_pcf_ert_ssl_cert=$(echo ${pcf_ert_ssl_cert} | sed 's/\s\+/\\\\r\\\\n/g' | sed
 my_pcf_ert_ssl_key=$(echo ${pcf_ert_ssl_key} | sed 's/\s\+/\\\\r\\\\n/g' | sed 's/\\\\r\\\\nRSA\\\\r\\\\nPRIVATE\\\\r\\\\nKEY/ RSA PRIVATE KEY/g')
 perl -pi -e "s|{{pcf_ert_ssl_cert}}|${my_pcf_ert_ssl_cert}|g" ${json_file}
 perl -pi -e "s|{{pcf_ert_ssl_key}}|${my_pcf_ert_ssl_key}|g" ${json_file}
-
-if [[ ! -f ${json_file} ]]; then
-  echo "Error: cant find file=[${json_file}]"
-  exit 1
-fi
 
 function fn_om_linux_curl_fail {
     echo ERROR
